Improved types
